Add an example for Grid: table
